+++ CHANGELOG.md
@@ -31,6 +31,7 @@ All notable changes to this project will be documented in this file.
 - DataGrid: Hide selection when disabled #178
 - DataGrid: Exception in AddDisplayControl #181
 - DataGrid: Support for custom type descriptors #200
+- DataGrid: Auto width not working with horizontal stretch #223
 - PropertyGrid: Support BrowsableAttribute on enum items when shown as ComboBox or ListBox #133
 - PropertyGrid: Hidden error info when HeaderPlacement = Above #212
 - TreeListBox: Applied workaround for "Height must be non-negative" exception #38 #142