+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Support for .NET Core 3.0
 
+### Fixed
+- PropertyGrid: support for INotifyDataErrorInfo #226
+
 ## [3.0.0]
 ### Changed
 - DataGrid: Commands moved to DataGridCommands